Рефакторинг + Приборка в тестах
