--- src/adapters/movies.js
@@ -1,235 +1,0 @@
-export const createMovie = (data) => {
-  return {
-    id: data.id,
-    title: data.name,
-    genre: data.genre,
-    year: data.released,
-    runTimeInSec: data.run_time * 60,
-    poster: {
-      image: data.poster_image,
-      title: ``,
-    },
-    bgPoster: {
-      image: data.background_image,
-      title: ``,
-    },
-    rating: {
-      score: data.rating,
-      count: data.scores_count
-    },
-    description: data.description,
-    director: data.director,
-    starrings: data.starring,
-    preview: data.preview_video_link,
-  };
-};
-
-export const movies = [
-  {
-    id: 1,
-    title: `Fantastic Beasts: The Crimes of Grindelwald`,
-    genre: `Documentary`,
-    year: 2011,
-    runTimeInSec: 90 * 60,
-    poster: {
-      image: `img/fantastic-beasts-the-crimes-of-grindelwald.jpg`,
-      title: `Fantastic Beasts: The Crimes of Grindelwald`,
-    },
-    bgPoster: {
-      image: `img/fantastic-beasts-the-crimes-of-grindelwald.jpg`,
-      title: `Fantastic Beasts: The Crimes of Grindelwald`,
-    },
-    rating: {
-      score: 3.0,
-      count: 48
-    },
-    description: `Description of Fantastic Beasts: The Crimes of Grindelwald.`,
-    director: `Director of Fantastic Beasts: The Crimes of Grindelwald.`,
-    starrings: [`Starring Star1`, `Starring Star2`, `Starring Star3`, `Starring Star4`, `Starring Star5`, `Starring Star6`, `Starring Star7`, `Starring Star8`],
-    preview: `https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4`,
-  },
-  {
-    id: 2,
-    title: `Bohemian Rhapsody`,
-    genre: `Documentary`,
-    year: 2012,
-    runTimeInSec: 120 * 60,
-    poster: {
-      image: `img/bohemian-rhapsody.jpg`,
-      title: `Bohemian Rhapsody`,
-    },
-    bgPoster: {
-      image: `img/bohemian-rhapsody.jpg`,
-      title: `Bohemian Rhapsody`,
-    },
-    rating: {
-      score: 10.0,
-      count: 171
-    },
-    description: `Description of Bohemian Rhapsody.`,
-    director: `Director of Bohemian Rhapsody.`,
-    starrings: [`Starring Star1`, `Starring Star2`, `Starring Star3`, `Starring Star4`, `Starring Star5`, `Starring Star6`, `Starring Star7`],
-    preview: `https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4`,
-  },
-  {
-    id: 3,
-    title: `Macbeth`,
-    genre: `Documentary`,
-    year: 2013,
-    runTimeInSec: 165 * 60,
-    poster: {
-      image: `img/macbeth.jpg`,
-      title: `Macbeth`,
-    },
-    bgPoster: {
-      image: `img/macbeth.jpg`,
-      title: `Macbeth`,
-    },
-    rating: {
-      score: 9.2,
-      count: 169
-    },
-    description: `Description of Macbeth.`,
-    director: `Director of Macbeth.`,
-    starrings: [`Starring Star1`, `Starring Star2`, `Starring Star3`, `Starring Star4`, `Starring Star5`, `Starring Star6`],
-    preview: `https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4`,
-  },
-  {
-    id: 4,
-    title: `Aviator`,
-    genre: `Documentary`,
-    year: 2014,
-    runTimeInSec: 58 * 60,
-    poster: {
-      image: `img/aviator.jpg`,
-      title: `Aviator`,
-    },
-    bgPoster: {
-      image: `img/aviator.jpg`,
-      title: `Aviator`,
-    },
-    rating: {
-      score: 7.3,
-      count: 217
-    },
-    description: `Description of Aviator.`,
-    director: `Director of Aviator.`,
-    starrings: [`Starring Star1`, `Starring Star2`, `Starring Star3`, `Starring Star4`, `Starring Star5`],
-    preview: `https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4`,
-  },
-  {
-    id: 5,
-    title: `We need to talk about Kevin`,
-    genre: `Documentary`,
-    year: 2015,
-    runTimeInSec: 143 * 60,
-    poster: {
-      image: `img/we-need-to-talk-about-kevin.jpg`,
-      title: `We need to talk about Kevin`,
-    },
-    bgPoster: {
-      image: `img/we-need-to-talk-about-kevin.jpg`,
-      title: `We need to talk about Kevin`,
-    },
-    rating: {
-      score: 4.5,
-      count: 125
-    },
-    description: `Description of We need to talk about Kevin.`,
-    director: `Director of We need to talk about Kevin.`,
-    starrings: [`Starring Star1`, `Starring Star2`, `Starring Star3`, `Starring Star4`],
-    preview: `https://upload.wikimedia.org/wikipedia/commons/transcoded/b/b3/Big_Buck_Bunny_Trailer_400p.ogv/Big_Buck_Bunny_Trailer_400p.ogv.360p.webm`,
-  },
-  {
-    id: 6,
-    title: `What We Do in the Shadows`,
-    genre: `Sci-Fi`,
-    year: 2016,
-    runTimeInSec: 202 * 60,
-    poster: {
-      image: `img/what-we-do-in-the-shadows.jpg`,
-      title: `What We Do in the Shadows`,
-    },
-    bgPoster: {
-      image: `img/what-we-do-in-the-shadows.jpg`,
-      title: `What We Do in the Shadows`,
-    },
-    rating: {
-      score: 8.1,
-      count: 321
-    },
-    description: `Description of What We Do in the Shadows.`,
-    director: `Director of What We Do in the Shadows.`,
-    starrings: [`Starring Star1`, `Starring Star2`, `Starring Star3`],
-    preview: `https://upload.wikimedia.org/wikipedia/commons/transcoded/b/b3/Big_Buck_Bunny_Trailer_400p.ogv/Big_Buck_Bunny_Trailer_400p.ogv.360p.webm`,
-  },
-  {
-    id: 7,
-    title: `Revenant`,
-    genre: `THrillers`,
-    year: 2017,
-    runTimeInSec: 99 * 60,
-    poster: {
-      image: `img/revenant.jpg`,
-      title: `Revenant`,
-    },
-    bgPoster: {
-      image: `img/revenant.jpg`,
-      title: `Revenant`,
-    },
-    rating: {
-      score: 1.2,
-      count: 27
-    },
-    description: `Description of Revenant.`,
-    director: `Director of Revenant.`,
-    starrings: [`Starring Star1`, `Starring Star2`],
-    preview: `https://upload.wikimedia.org/wikipedia/commons/transcoded/b/b3/Big_Buck_Bunny_Trailer_400p.ogv/Big_Buck_Bunny_Trailer_400p.ogv.360p.webm`,
-  },
-  {
-    id: 8,
-    title: `Johnny English`,
-    genre: `Comedies`,
-    year: 2018,
-    runTimeInSec: 174 * 60,
-    poster: {
-      image: `img/johnny-english.jpg`,
-      title: `Johnny English`,
-    },
-    bgPoster: {
-      image: `img/johnny-english.jpg`,
-      title: `Johnny English`,
-    },
-    rating: {
-      score: 6.4,
-      count: 134
-    },
-    description: `Description of Johnny English.`,
-    director: `Director of Johnny English.`,
-    starrings: [`Starring Star1`],
-    preview: `https://upload.wikimedia.org/wikipedia/commons/transcoded/b/b3/Big_Buck_Bunny_Trailer_400p.ogv/Big_Buck_Bunny_Trailer_400p.ogv.360p.webm`,
-  },
-  {
-    id: 9,
-    title: `Orlando`,
-    genre: `Fantasy`,
-    year: 2019,
-    runTimeInSec: 115 * 60,
-    poster: {
-      image: `img/orlando.jpg`,
-      title: `Orlando`,
-    },
-    bgPoster: {
-      image: `img/orlando.jpg`,
-      title: `Orlando`,
-    },
-    rating: {
-      score: 7.7,
-      count: 67
-    },
-    description: `Description of Orlando.`,
-    director: `Director of Orlando.`,
-    starrings: [`Starring Star1`, `Starring Star9`],
-    preview: `https://upload.wikimedia.org/wikipedia/commons/transcoded/b/b3/Big_Buck_Bunny_Trailer_400p.ogv/Big_Buck_Bunny_Trailer_400p.ogv.360p.webm`,
-  },
-];
