--- src/adapters/promo-movie-reviews.js
@@ -1,30 +1,0 @@
-// export const promoMovieReviews = [
-//   {
-//     movieId: 555,
-//     author: `Reviewer-1`,
-//     dateUTC: new Date(4586616320000),
-//     ratingScore: 1.0,
-//     comment: `Comment Author-1 to Movie-1.`,
-//   },
-//   {
-//     movieId: 555,
-//     author: `Reviewer-2`,
-//     dateUTC: new Date(6586616320000),
-//     ratingScore: 2.0,
-//     comment: `Comment Author-2 to Movie-2.`,
-//   },
-//   {
-//     movieId: 555,
-//     author: `Reviewer-3`,
-//     dateUTC: new Date(7586616320000),
-//     ratingScore: 3.0,
-//     comment: `Comment Author-3 to Movie-3.`,
-//   },
-//   {
-//     movieId: 555,
-//     author: `Reviewer-4`,
-//     dateUTC: new Date(8586616320000),
-//     ratingScore: 4.1,
-//     comment: `Comment Author-3 to Movie-4.`,
-//   },
-// ];
