--- src/adapters/promo-movie.js
@@ -1,24 +1,0 @@
-export const promoMovie = {
-  id: 555,
-  title: `The Grand Budapest Hotel`,
-  genre: `Drama`,
-  year: 2014,
-  runTimeInSec: 90 * 60,
-  poster: {
-    image: `img/macbeth.jpg`,
-    title: `The Grand Budapest Hotel poster`,
-  },
-  bgPoster: {
-    image: `img/bg-the-grand-budapest-hotel.jpg`,
-    title: `The Grand Budapest Hotel`,
-  },
-  rating: {
-    score: 8.9,
-    count: 240
-  },
-  description: `In the 1930s, the Grand Budapest Hotel is a popular European ski resort, presided over by concierge Gustave H. (Ralph Fiennes). Zero, a junior lobby boy, becomes Gustave's friend and protege.
-  Gustave prides himself on providing first-class service to the hotel's guests, including satisfying the sexual needs of the many elderly women who stay there. When one of Gustave's lovers dies mysteriously, Gustave finds himself the recipient of a priceless painting and the chief suspect in her murder.`,
-  director: `Wes Andreson`,
-  starrings: [`Bill Murray`, `Edward Norton`, `Jude Law`, `Willem Dafoe`, `Saoirse Ronan`, `Tony Revoloru`, `Tilda Swinton`],
-  preview: `https://upload.wikimedia.org/wikipedia/commons/transcoded/b/b3/Big_Buck_Bunny_Trailer_400p.ogv/Big_Buck_Bunny_Trailer_400p.ogv.360p.webm`,
-};
